Added Paul's Work Instructions

+++ DAQRI Headset Repair Project/Assets/Assets/DAQRI/System/Scripts/Sensors/CameraDevice.cs
@@ -339,7 +339,7 @@ namespace DAQRI {
 		}
 
 		private void calculateAspectRatio(){
-			aspectRatio = width / height;
+            aspectRatio = (float)width / height;
 		}
 
 		private void updateCameraRawData()
